requirements.txt
requests
bitcoinlib
